Splitting chat sessions class
