+++ src/vs/workbench/contrib/chat/browser/actions/chatSessionActions.ts
@@ -3,20 +3,26 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
+import * as nls from '../../../../../nls.js';
 import { Codicon } from '../../../../../base/common/codicons.js';
 import { KeyCode } from '../../../../../base/common/keyCodes.js';
 import { MarshalledId } from '../../../../../base/common/marshallingIds.js';
+import { IChatSessionRecommendation } from '../../../../../base/common/product.js';
 import Severity from '../../../../../base/common/severity.js';
 import { localize } from '../../../../../nls.js';
 import { Action2, MenuId, MenuRegistry } from '../../../../../platform/actions/common/actions.js';
 import { IConfigurationService } from '../../../../../platform/configuration/common/configuration.js';
 import { ContextKeyExpr } from '../../../../../platform/contextkey/common/contextkey.js';
 import { IDialogService } from '../../../../../platform/dialogs/common/dialogs.js';
+import { IExtensionGalleryService } from '../../../../../platform/extensionManagement/common/extensionManagement.js';
 import { ServicesAccessor } from '../../../../../platform/instantiation/common/instantiation.js';
 import { KeybindingWeight } from '../../../../../platform/keybinding/common/keybindingsRegistry.js';
 import { ILogService } from '../../../../../platform/log/common/log.js';
+import { IProductService } from '../../../../../platform/product/common/productService.js';
+import { IQuickInputService } from '../../../../../platform/quickinput/common/quickInput.js';
 import { IEditorGroupsService } from '../../../../services/editor/common/editorGroupsService.js';
 import { AUX_WINDOW_GROUP, IEditorService, SIDE_GROUP } from '../../../../services/editor/common/editorService.js';
+import { IWorkbenchExtensionManagementService } from '../../../../services/extensionManagement/common/extensionManagement.js';
 import { IViewsService } from '../../../../services/views/common/viewsService.js';
 import { ChatContextKeys } from '../../common/chatContextKeys.js';
 import { IChatService } from '../../common/chatService.js';
@@ -26,10 +32,11 @@ import { ChatConfiguration } from '../../common/constants.js';
 import { ChatViewId, IChatWidgetService } from '../chat.js';
 import { IChatEditorOptions } from '../chatEditor.js';
 import { ChatEditorInput } from '../chatEditorInput.js';
-import { VIEWLET_ID } from '../chatSessions.js';
 import { ChatSessionItemWithProvider, findExistingChatEditorByUri, isLocalChatSessionItem } from '../chatSessions/common.js';
 import { ChatViewPane } from '../chatViewPane.js';
 import { CHAT_CATEGORY } from './chatActions.js';
+import { CancellationToken } from '../../../../../base/common/cancellation.js';
+import { VIEWLET_ID } from '../chatSessions/view/chatSessionsView.js';
 
 export interface IChatSessionContext {
 	sessionId: string;
@@ -358,6 +365,65 @@ export class ToggleChatSessionsDescriptionDisplayAction extends Action2 {
 	}
 }
 
+export class ChatSessionsGettingStartedAction extends Action2 {
+	static readonly ID = 'chat.sessions.gettingStarted';
+
+	constructor() {
+		super({
+			id: ChatSessionsGettingStartedAction.ID,
+			title: nls.localize2('chat.sessions.gettingStarted.action', "Getting Started with Chat Sessions"),
+			icon: Codicon.sendToRemoteAgent,
+			f1: false,
+		});
+	}
+
+	override async run(accessor: ServicesAccessor): Promise<void> {
+		const productService = accessor.get(IProductService);
+		const quickInputService = accessor.get(IQuickInputService);
+		const extensionManagementService = accessor.get(IWorkbenchExtensionManagementService);
+		const extensionGalleryService = accessor.get(IExtensionGalleryService);
+
+		const recommendations = productService.chatSessionRecommendations;
+		if (!recommendations || recommendations.length === 0) {
+			return;
+		}
+
+		const installedExtensions = await extensionManagementService.getInstalled();
+		const isExtensionAlreadyInstalled = (extensionId: string) => {
+			return installedExtensions.find(installed => installed.identifier.id === extensionId);
+		};
+
+		const quickPickItems = recommendations.map((recommendation: IChatSessionRecommendation) => {
+			const extensionInstalled = !!isExtensionAlreadyInstalled(recommendation.extensionId);
+			return {
+				label: recommendation.displayName,
+				description: recommendation.description,
+				detail: extensionInstalled
+					? nls.localize('chatSessions.extensionAlreadyInstalled', "'{0}' is already installed", recommendation.extensionName)
+					: nls.localize('chatSessions.installExtension', "Installs '{0}'", recommendation.extensionName),
+				extensionId: recommendation.extensionId,
+				disabled: extensionInstalled,
+			};
+		});
+
+		const selected = await quickInputService.pick(quickPickItems, {
+			title: nls.localize('chatSessions.selectExtension', "Install Chat Extensions"),
+			placeHolder: nls.localize('chatSessions.pickPlaceholder', "Choose extensions to enhance your chat experience"),
+			canPickMany: true,
+		});
+
+		if (!selected) {
+			return;
+		}
+
+		const galleryExtensions = await extensionGalleryService.getExtensions(selected.map(item => ({ id: item.extensionId })), CancellationToken.None);
+		if (!galleryExtensions) {
+			return;
+		}
+		await extensionManagementService.installGalleryExtensions(galleryExtensions.map(extension => ({ extension, options: { preRelease: productService.quality !== 'stable' } })));
+	}
+}
+
 // Register the menu item - show for all local chat sessions (including history items)
 MenuRegistry.appendMenuItem(MenuId.ChatSessionsMenu, {
 	command: {
@@ -422,4 +488,15 @@ MenuRegistry.appendMenuItem(MenuId.ViewContainerTitle, {
 	group: '1_config',
 	order: 1,
 	when: ContextKeyExpr.equals('viewContainer', VIEWLET_ID),
+});
+
+MenuRegistry.appendMenuItem(MenuId.ViewTitle, {
+	command: {
+		id: 'workbench.action.openChat',
+		title: nls.localize2('interactiveSession.open', "New Chat Editor"),
+		icon: Codicon.plus
+	},
+	group: 'navigation',
+	order: 1,
+	when: ContextKeyExpr.equals('view', `${VIEWLET_ID}.local`),
 });
